+++ hut_10sqft/package.xml
@@ -7,7 +7,6 @@
   <url type="bugtracker">https://github.com/130s/hut_10sqft/issues</url>
   <author>Isaac Saito</author>
 
-  <buildtool_depend>ament</buildtool_depend>
   <exec_depend>ack</exec_depend>
   <exec_depend>apt-transport-https</exec_depend>
   <exec_depend>ca-certificates</exec_depend>
@@ -19,6 +18,7 @@
   <!-- <exec_depend>gitk</exec_depend> Issue to be resolved https://github.com/kinu-garage/hut_10sqft/pull/1081#issuecomment-2334335301 -->
   <exec_depend>gnuplot</exec_depend>
   <exec_depend>gnuplot-x11</exec_depend>
+  <exec_depend>hut_10sqft_lib</exec_depend>
   <exec_depend>iftop</exec_depend>
   <exec_depend>imagemagick</exec_depend>
   <exec_depend>ipython3</exec_depend>
@@ -48,4 +48,11 @@
   <exec_depend>tree</exec_depend>
   <exec_depend>xdotool</exec_depend>  # https://github.com/kinu-garage/hut_10sqft/issues/1077
 
+  <test_depend>ament_copyright</test_depend>
+  <test_depend>ament_flake8</test_depend>
+  <test_depend>ament_pep257</test_depend>
+  <test_depend>python3-pytest</test_depend>
+  <export>
+    <build_type>ament_python</build_type>
+  </export>
 </package>
