--- hut_10sqft/pyproject.toml
@@ -1,35 +1,0 @@
-[build-system]
-requires      = ["setuptools>=61.0.0", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "hut_10sqft"
-version = "0.1.0"
-description = "Operating system setup for personal use"
-readme = "README.md"
-authors = [{ name = "Isaac Saito", email = "iisaac.saito@gmail.com" }]
-license = { file = "LICENSE" }
-classifiers = [
-    "License :: OSI Approved :: Apache2.0",
-    "Programming Language :: Python",
-    "Programming Language :: Python :: 3",
-]
-keywords = ["OS setup for personal usecases"]
-dependencies = [
-    # "openpyxl>=3.0.6",  # TODO Consider https://github.com/kinu-garage/nton_matching/issues/19
-    "rosdep"
-]
-requires-python = ">=3.10"
-
-[project.optional-dependencies]
-dev = ["ipython", "pip-tools", "pytest"]
-
-[project.urls]
-Homepage = "https://github.com/kinu-garage/hut_10sqft.git"
-
-[project.scripts]
-suco = "hut_10sqft.__main__:main"
-
-[tool.setuptools]
-package-dir = {"" = "src"}
-py-modules = ["src"]
