--- hut_10sqft/src/hut_10sqft/__main__.py
@@ -1,13 +1,0 @@
-import sys
-
-from hut_10sqft.init_setup import CompInitSetup
-
-
-def main():
-    # Check Python environment
-    print("Python sys.path: {}".format(sys.path))
-    comp_setup = CompInitSetup()
-    comp_setup.main()
-
-if __name__ == "__main__":
-    main()
