--- hut_10sqft/src/hut_10sqft/init_setup.py
@@ -1,1583 +1,0 @@
-#! /usr/bin/env python3
-
-# Copyright (C) 2023 Kinu Garage
-# Licensed under Apache 2
-
-import argparse
-try:
-    import apt
-    import pkg_resources
-except ModuleNotFoundError as e:
-    print(f"This module isn't available at the moment but will be installed later.\n{str(e)}")
-try:
-    import git
-except ModuleNotFoundError as e:
-    print(f"This module isn't available at the moment but will be installed later.\n{str(e)}")
-import importlib
-import logging
-import os
-import pathlib
-import platform
-import pwd
-import shlex
-import shutil
-import subprocess
-import sys
-from typing import List
-from datetime import datetime
-
-
-class HostConf():
-    def __init__(self, hostname, bash_cfg, emacs_cfg, sshkey_prv, sshkey_pub):
-        self._hostname = hostname
-        self._bash_cfg = bash_cfg
-        self._emacs_cfg = emacs_cfg
-        self._sshkey_prv = sshkey_prv
-        self._sshkey_pub = sshkey_pub
-
-    @property
-    def hostname(self) -> str:
-        return self._hostname
-
-    @hostname.setter
-    def hostname(self, v: str):
-        self._hostname = v
-
-    @property
-    def bash_cfg(self):
-        return self._bash_cfg
-
-    @property
-    def emacs_cfg(self):
-        return self._emacs_cfg
-
-    @property
-    def sshkey_prv(self):
-        return self._sshkey_prv
-
-    @property
-    def sshkey_pub(self):
-        return self._sshkey_pub
-
-
-class ConfigDispach():
-    """Need for the setter of each entry is questionable in the beginning though"""
-    def __init__(self, path_source, path_dest=None, is_symlink=False, necessary=True, hint_enable=""):
-        """
-        @param necessary: If True, when the path in `path_source` not found then the accessor should raise an error.
-        @param hint_enable: Hint to enable the symlink creation, e.g. how to set up a mount on to Google Drive directory from Linux container on ChromeOS.
-        """
-        self._path_source = path_source
-        self._path_dest = path_dest
-        self._is_symlink = is_symlink
-        self._necessary = necessary
-        self._hint_enable = hint_enable
-
-    @property
-    def path_source(self):
-        return self._path_source
-
-    @path_source.setter
-    def path_source(self, v):
-        self._path_source = v
-
-    @property
-    def path_dest(self):
-        return self._path_dest
-
-    @path_dest.setter
-    def path_dest(self, v):
-        self._path_dest = v
-
-    @property
-    def is_symlink(self):
-        return self._is_symlink
-
-    @is_symlink.setter
-    def is_symlink(self, v):
-        """Must be absolute path"""
-        self._is_symlink = v
-
-    @property
-    def necessary(self) -> bool:
-        return self._necessary
-
-    @necessary.setter
-    def necessary(self, v: bool):
-        self._necessary = v
-
-    @property
-    def hint_enable(self) -> str:
-        return self._hint_enable
-
-
-class OsUtil:
-    """
-    @summary: Utility for Operating System handling.
-    """
-    SUFFIX_BACKUP = ".bk"
-    _LOGGER_NAME = "OsUtil-logger"
-    _MSG_EXEC_NOT_FOUND = "Cannot find '{}' executable."
-
-    ERRORCOCDE_APTCACHE_NOCANDIDATE = -1001
-    ERRORCOCDE_APTCACHE_CANDIDATE_NOTINSTALLED = -1002
-
-    TYPE_OS_CHROMEOS = "ChromeOS"
-    TYPE_OS_LINUX = "Linux"
-    TYPE_OS_MACOS = "MacOS"
-    TYPE_OS_MACOS_DARWIN = "Darwin"
-    TYPE_LINUX_DISTRO_DEBIAN = "Debian"
-    TYPE_LINUX_DISTRO_UBUNTU = "Ubuntu"
-
-    def __init__(self, logger=None):
-        if logger:
-            self._logger = logger
-        else:
-            self._logger = OsUtil._gen_logger()
-
-    @staticmethod
-    def _gen_logger(logger_name=_LOGGER_NAME, log_level=logging.DEBUG):
-        logger = logging.getLogger(logger_name)
-        log_handler = logging.StreamHandler()        
-        logger.setLevel(log_level)
-        if not logger.hasHandlers():
-            logger.addHandler(log_handler)
-        return logger
-
-    @staticmethod
-    def setup_rosdep():
-        OsUtil.subproc_bash(f"rosdep init", does_sudo=True)
-        OsUtil.subproc_bash(f"rosdep update")
-        OsUtil.subproc_bash(f"apt update", does_sudo=True)
-
-    @staticmethod
-    def apt_install(deb_pkg_names: list[str], logger=None):
-        """
-        @exception LookupError: If 'apt' executable is not found on the OS.
-        """
-        if not logger:
-            logger = OsUtil._gen_logger()
-        logger.info(f"Installing by apt: {deb_pkg_names}")
-        OsUtil._apt_install_bash(deb_pkg_names, logger)
-
-    @staticmethod
-    def _apt_cache_policy(debpkg_name: str, logger=None):
-        """
-        @rtype: str, str, int
-        @return: This method returns 3 things and by default they are what Python's subprocess returns. However, when the following conditions are met, this method might overwrite 'error' and ret_code' before returning.
-          - When a candidate for 'debpkg_name' not found, ret_code = -1001
-          - When a candidate found but not found installed, ret_code = -1002
-        """
-        output, error, ret_code = OsUtil.subproc_bash(f"apt-cache policy {debpkg_name}")
-        if not output:
-            ret_code = OsUtil.ERRORCOCDE_APTCACHE_NOCANDIDATE
-            error = f"'{debpkg_name}' is not found on this OS, even as an install candidate. Check if it is really available."
-        elif f"{debpkg_name}:\n  Installed: (none)" in output:
-            ret_code = OsUtil.ERRORCOCDE_APTCACHE_CANDIDATE_NOTINSTALLED
-            error = f"The install candidate of the pkg '{debpkg_name}' found on this OS, but 'apt-cache policy' didn't find it installed."
-        return output, error, ret_code
-        
-    @staticmethod
-    def apt_cache_policy(debpkg_names: list[str], logger=None):
-        """
-        @brief: Prints the apt-cache policy for the given deb package names.
-        @param debpkg_names: List of deb package names to check.
-        @exception RuntimeWarning: When one or more pkgs found not installed.
-        """
-        if not logger:
-            logger = OsUtil._gen_logger()
-        if not debpkg_names:
-            raise ValueError("No deb package names passed to 'apt_cache_policy' method.")
-        if (type(debpkg_names) is not list):
-            debpkg_names = OsUtil._encapsulate_if_string(debpkg_names)
-        if " " in debpkg_names:
-            raise ValueError(f"Space found in the input that is supposed to be a list of pkg names: {debpkg_names}")
-
-        errors = []
-        pkgs_success = []
-        for pkg_name in debpkg_names:
-            output, error, ret_code = OsUtil._apt_cache_policy(pkg_name)
-            if ret_code != 0:
-                errors.append(f"Failed to get apt-cache policy for '{pkg_name}'. Error: {error}")
-            else:
-                logger.info(f"'apt-cache policy' result for '{pkg_name}':\n\t{output}")
-        _msg_result_header = f"Report: Package installation status:\n"
-        _msg_result_header_success = f"- Packages found installed: {pkgs_success}"
-        _msg_all = f"{_msg_result_header}\n\t{_msg_result_header_success}"
-        if errors:
-            _str_errors = ""
-            for error in errors:
-                _str_errors += "\t" + error + "\n"
-            raise RuntimeWarning(f"{_msg_all}\n\t- The following pkgs didn't get installed: {_str_errors}")
-        return _msg_all
-
-    @staticmethod
-    def _encapsulate_if_string(var) -> List[str]:
-        """
-        @summary: If 'var' is a string, encapsulate it in a list.
-        """
-        if (type(var) is str) and (" " not in var):
-            var = [var]
-        return var
-
-    @staticmethod
-    def _apt_install_bash(deb_pkgs_name: list[str], logger=None):
-        if (type(deb_pkgs_name) is not list):
-            deb_pkgs_name = OsUtil._encapsulate_if_string(deb_pkgs_name)
-
-        # 'deb_pkgs_name' is a list of strings, while subprocess takes an input literally
-        # so if a list is spplied then it'd take square brackets and would return an error.
-        # Thus need to expand as a non-list, single string.
-        deb_pkg_names_str = " ".join(deb_pkgs_name)
-
-        OsUtil.subproc_bash(f"apt update", does_sudo=True)
-        OsUtil.subproc_bash(f"DEBIAN_FRONTEND=noninteractive apt install -y {deb_pkg_names_str}", does_sudo=True)
-        # Just to verify, print 'apt-cache policy' output for the 'deb_pkg_names_str'.
-        OsUtil.apt_cache_policy(deb_pkgs_name)
-
-    @staticmethod
-    def _apt_install_py(deb_pkg_name, logger=None):
-        """
-        @deprecated: Unsure if this method should be deprecated but it doesn't seem to be used.
-        """
-        cache = apt.cache.Cache()
-        cache.update()
-        cache.open()
-
-        pkg = cache[deb_pkg_name]
-        if pkg.is_installed:
-            logger.info("{pkg_name} already installed".format(pkg_name=deb_pkg_name))
-        else:
-            pkg.mark_install()
-
-        try:
-            cache.commit()
-        except SystemError as e:
-            logger.error(sys.stderr, "Sorry, package installation failed [{err}]".format(err=str(e)))
-
-    @staticmethod
-    def install_pip_adhoc(pip_pkgs=[], logger=None, allow_break=False):
-        """
-        @param allow_break: If True, `pip` runs with '--break-system-packages' option.
-        """
-        if not logger:
-            logger = OsUtil._gen_logger()
-        if not pip_pkgs:
-            logger.warning(f"No pip pkgs requested to be installed, so skpping. Passed: {pip_pkgs}")
-            return
-        _PIP_OPTION_BREAK = "--break-system-packages"
-        cmd_list = ['pip', 'install', *pip_pkgs]
-        
-        if allow_break:
-            cmd_list.append(_PIP_OPTION_BREAK)
-            _msg = f"Cmd: {cmd_list}."            
-            _msg = _msg + " See https://stackoverflow.com/questions/75602063 for the risk of passing '--break-system-packages' option."
-        else:
-            _msg = f"Cmd: {cmd_list}."
-        logger.info(_msg)
-        _attempts = 0
-        _ret_code = -1
-        while (not _ret_code) and (_attempts < 3):
-            try:
-                _ret_code = subprocess.check_call(cmd_list)
-            except subprocess.CalledProcessError as e:
-                _err_msg = f"{_attempts}-th 'pip install' attempt failed. Often network issue. Re-trying. Error: {str(e)}"
-                if allow_break:
-                    _err_msg += f"\n\tAnother possible reason is 'pip' is too old so that an option {_PIP_OPTION_BREAK} is not available (see https://stackoverflow.com/a/78600962/577001). Re-trying without that."
-                    cmd_list.pop()  # Popping the last element is not super robust way to get rid of '_PIP_OPTION_BREAK'...
-                logger.error(_err_msg)
-
-            _attempts += 1        
-
-    @staticmethod
-    def copy_prop_file(path_src, path_dest):
-        """
-        @brief: Python's file copy methods are known to be missing an option
-            to copy meta data. This method copies them from a file to another.
-        @return: True if copying metadata was successful.
-        @raise AssertionError: When either UID / GID / file size / st_dev
-            differs b/w src and dest files.
-        """
-        meta_src = os.stat(path_src)
-        os.chown(path_dest, meta_src.st_uid, meta_src.st_gid)
-
-        meta_dest = os.stat(path_dest)
-        if (meta_src.st_uid != meta_dest.st_uid) or \
-           (meta_src.st_gid != meta_dest.st_gid) or \
-           (meta_src.st_dev != meta_dest.st_dev) or \
-           (meta_src.st_size != meta_dest.st_size):
-            raise AssertionError("Copying meta data failed. Metadata per file:\n\tSrc: {}\n\tDst: {}".format(meta_src, meta_dest))
-        else:
-            return True
-
-    @staticmethod
-    def subproc_bash(
-            cmd,
-            does_sudo=False,
-            print_stdout_err=False,
-            logger=None,
-            non_interactive=False):
-        if not logger:
-            logger = OsUtil._gen_logger()  
-        if not cmd:
-            raise ValueError("Command to execute not passed.")
-
-        bash_type = '/bin/sh'
-        bash_arg = '-c'
-        bash_full_cmd = [bash_type, bash_arg]
-        if does_sudo == True:
-            bash_full_cmd.insert(0, 'sudo')
-
-        if non_interactive:
-            cmd = "DEBIAN_FRONTEND=noninteractive " + cmd
-        bash_full_cmd.append(cmd)
-
-        logger.info(f"subprocess: About to execute the cmd: {bash_full_cmd}")
-        _subproc = None
-        if print_stdout_err:
-            _subproc = subprocess.Popen(bash_full_cmd)
-        else:
-            while not _subproc:  # TODO Afraid this look could lead an infinite loop.
-                try:
-                    _subproc = subprocess.Popen(bash_full_cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-                except FileNotFoundError as e:
-                    del bash_full_cmd[0]
-                    logger.warning(f"If 'sudo' is not found on this env, remove that from the command set. New command: {bash_full_cmd}. Retry now.")
-
-        output, error = _subproc.communicate()
-        bash_return_code = _subproc.returncode
-
-        if (output or error) is not None:
-            try:
-                output = output.decode("utf-8").rstrip('\n')
-                error = error.decode("utf-8").rstrip('\n')
-            except UnicodeDecodeError:
-                _ERR_MSG = "Potentially 'UnicodeDecodeError'"
-                output = _ERR_MSG
-                error = _ERR_MSG
-        logger.info(f"{bash_return_code=}, {output=}, {error=}")
-        return output, error, bash_return_code
-
-    @staticmethod
-    def create_parent_dir(path_dest: str, logger=None):
-        if not logger:
-            logger = OsUtil._gen_logger()         
-        path_dir_dest = pathlib.Path(path_dest).parent
-        logger.info(f"If the directory of the target for {path_dest} doesn't exist (i.e. {path_dir_dest}), create it.")
-        if not os.path.exists(path_dir_dest):
-            os.mkdir(path_dir_dest)
-
-    @staticmethod
-    def copy_a_file(path_source: str, path_dest: str, is_symlink=False, overwrite=False, backup_suffix=".org", logger=None):
-        """
-        @summary A tool to take the list of conf files, place them at the designated location so that each application can find them.
-        @param backup_suffix: Only used when 'overwrite' is True, NOTE if no string is passed, the original dest file will be DELETED.
-        @return: 
-        - True if dest exists after the process.
-        - Timestamp of the file copied.
-        @todo Remove dependency on ConfigDispach. This method can be written with just taking str.
-        """
-        if not logger:
-            logger = OsUtil._gen_logger()
-        logger.debug("poku.path_dest: {}".format(path_dest))
-        # Screening
-        if (not overwrite) and pathlib.Path(path_dest).exists():
-            raise FileExistsError("'{}' already exists.".format(path_dest))
-        if not os.path.exists(path_source):
-            raise FileNotFoundError("Source file '{}' not found.".format(path_source))
-
-        # If one direct parent folder for the destination doesn't exist, create one.
-        OsUtil.create_parent_dir(path_dest)
-
-        _timestamp = datetime.today().strftime("%Y%m%d-%H%M%S")
-        if overwrite:
-            if backup_suffix:
-                _backup_file_path = os.path.join(path_dest + "_" + _timestamp + backup_suffix)
-                shutil.copyfile(path_dest, _backup_file_path)
-                logger.info(f"File '{path_dest}' is backed up at '{_backup_file_path}'")
-            os.remove(path_dest)
-            logger.info(f"File '{path_dest} was deleted without backup per instruction.")
-
-        if is_symlink:
-            os.symlink(path_source, path_dest)
-            logger.info("Created symlink at {}".format(path_dest))
-        else:
-            shutil.copyfile(path_source, path_dest)
-            logger.info("Moved a file at {}".format(path_dest))
-
-        return pathlib.Path(path_dest).exists(), _timestamp
-
-    @staticmethod
-    def tilde_to_expand(value_to_scan, logger=None):
-        """
-        @type value_to_scan: str
-        @return: String after tilde-to-absolute path expansion.
-        @todo Entire input string is scanned and expanded, which may meet some usecases but other usecases may need something else.
-        """
-        val_result = value_to_scan
-        if not logger:
-            logger = OsUtil._gen_logger()
-        if (value_to_scan) and (value_to_scan.find("~") != -1):  # When v is not none and contains tilde
-            val_result = pathlib.Path(value_to_scan).expanduser()
-            logger.info(f"Expanding a path that contains tilde with user ID. BEFORE: '{value_to_scan}', AFTER: {val_result}")
-        return val_result
-
-    @staticmethod
-    def get_repo_basename_from_url(url: str) -> str:
-        """
-        @see https://stackoverflow.com/a/55137835/577001
-        """
-        last_slash_index = url.rfind("/")
-        last_suffix_index = url.rfind(".git")
-        if last_suffix_index < 0:
-            last_suffix_index = len(url)
-        if last_slash_index < 0 or last_suffix_index <= last_slash_index:
-            raise Exception("Badly formatted url {}".format(url))
-        return url[last_slash_index + 1:last_suffix_index]
-
-    @staticmethod
-    def which(executable_name: str, throw_exception=False) -> str:
-        """
-        @return: None when `throw_exception` is False AND the path to `executable_name` is not found.
-        @raise ReferenceError: When `executable_name` not available, only when `throw_exception` is True.
-        """
-        path = shutil.which(executable_name)
-        if (not path) and throw_exception:
-            raise ReferenceError(f"The executable '{executable_name}' not found.")
-        return path
-
-    @staticmethod
-    def get_os_type(logger=None) -> tuple[str, str]:
-        _FILEPATH_OS_RELEASE = "/etc/os-release"
-        _FILEPATH_LSB_RELEASE = "/etc/lsb-release"
-        _type_os = ""
-        _type_distro = "None"
-        if not logger:
-            logger = OsUtil._gen_logger()
-        if platform.system() == OsUtil.TYPE_OS_LINUX:
-            _type_os = OsUtil.TYPE_OS_LINUX
-            try:
-                with open(_FILEPATH_OS_RELEASE, "r") as f:
-                    content = f.read()
-                    if f"ID={OsUtil.TYPE_LINUX_DISTRO_DEBIAN.lower()}" in content or f"ID_LIKE={OsUtil.TYPE_LINUX_DISTRO_DEBIAN.lower()}" in content:
-                        _type_distro = OsUtil.TYPE_LINUX_DISTRO_DEBIAN
-                    elif f"ID={OsUtil.TYPE_LINUX_DISTRO_UBUNTU.lower()}" in content or f"ID_LIKE={OsUtil.TYPE_LINUX_DISTRO_UBUNTU.lower()}" in content:
-                        _type_distro = OsUtil.TYPE_LINUX_DISTRO_UBUNTU
-                    else:
-                        raise RuntimeError(f"Running on another Linux distribution SUCO does not support. Content of {_FILEPATH_OS_RELEASE}: {content}")
-            except FileNotFoundError as e:
-                try:
-                    with open(_FILEPATH_LSB_RELEASE, "r") as f:
-                        content = f.read()
-                        if f"DISTRIB_ID={OsUtil.TYPE_LINUX_DISTRO_DEBIAN}" in content:
-                            _type_distro = OsUtil.TYPE_LINUX_DISTRO_DEBIAN
-                        elif f"DISTRIB_ID={OsUtil.TYPE_LINUX_DISTRO_UBUNTU}" in content:
-                            _type_distro = OsUtil.TYPE_LINUX_DISTRO_UBUNTU
-                        else:
-                            raise RuntimeError(f"Running on another Linux distribution SUCO does not support. Content of {_FILEPATH_LSB_RELEASE}: {content}")
-                except FileNotFoundError:
-                    raise RuntimeError(f"Could not determine the Linux distribution type. Neither '{_FILEPATH_OS_RELEASE}' nor '{_FILEPATH_LSB_RELEASE}' found.")
-        elif (platform.system() == OsUtil.TYPE_OS_MACOS) or (platform.system() == OsUtil.TYPE_OS_MACOS_DARWIN):
-            _type_os = OsUtil.TYPE_OS_MACOS
-        if not _type_os:
-            raise RuntimeError(f"OS type undetected or unsupported type found: '{platform.system()}'")
-        return _type_os, _type_distro
-
-
-class AbstCompSetupFactory():
-    """
-    @description: Applyig Abstract Factory pattern.
-    """
-    _DIR_DROXBOX_CONTAINER = "data"  # This is beyond programming, something that sticks with 130s' computer usage for decades.
-
-    def __init__(self, os_name="", args_in: argparse.Namespace=None):
-        self._os = os_name
-        self.init_logger(logger_name=__name__)
-        self._list_runtime_issues = []
-
-        # Create a conf folder under ~/.
-        self._path_base_conf = os.path.join(pathlib.Path.home(), ".config")
-        if not os.path.exists(self._path_base_conf):
-            os.makedirs(self._path_base_conf)
-
-    @property
-    def list_runtime_issues(self):
-        return self._list_runtime_issues
-
-    @property
-    def path_base_conf(self):
-        return self._path_base_conf
-
-    @path_base_conf.setter
-    def path_base_conf(self, value):
-        self._path_base_conf = value
-
-    def add_runtime_issue(self, value):
-        """
-        @param value: Although the type of this not strictly enforced, recommended to be Exption type.
-        """
-        self._list_runtime_issues.append(value)
-
-    def listup_runtime_issues(self):
-        if not self.list_runtime_issues:
-            self._logger.info("No runtime issues recorded:")
-            return
-
-        self._logger.info("Runtime issues captured:")
-        _issue_count = 1
-        for issue in self.list_runtime_issues:
-            self._logger.warning(f"\tIssue #{_issue_count}: {str(issue)}")
-            _issue_count += 1
-
-    def init_logger(self, logger_name, logger_level=logging.DEBUG):
-        self._logger = logging.getLogger(logger_name)
-        log_handler = logging.StreamHandler()
-        self._logger.setLevel(logger_level)
-        self._logger.addHandler(log_handler)
-
-    def setup_ros_installer_src(self):
-        raise NotImplementedError()
-
-    def setup_git_config(self, path_local_perm_conf):
-        raise NotImplementedError()
-
-    def setup_dropbox(self):
-        raise NotImplementedError()
-
-    def swap_caps_ctrl(self):
-        raise NotImplementedError()
-
-    def update_hostname(self, hostname):
-        raise NotImplementedError("Updating hostname feature is not yet implemented.")
-
-    def __str__(self) -> str:
-        return f"{self.__class__.__name__}(os_name={self._os_name})"
-
-    def __repr__(self) -> str:
-        return f"{type(self).__name__}(os_name={self._os_name})"
-
-    def run(self, host_config, conf_repo_remote, conf_base_path=""):
-        raise NotImplementedError()
-
-    def generate_symlinks(self, rootpath_symlinks, path_user_home=""):
-        raise NotImplementedError()
-
-
-class WindowsSetup(AbstCompSetupFactory):
-    def __init__(self, os_name):
-        raise NotImplementedError()
-
-
-class ShellCapableOsSetup(AbstCompSetupFactory):
-    """
-    @summary: Operating system that is capable of bash, zsh or any *sh shell that meets this tool's requirement.
-    """
-    def __init__(self, os_name="", args_in: argparse.Namespace=None):
-        super().__init__(os_name, args_in)
-
-        self._args_in = args_in
-
-        if not args_in:  # Setting rather arbitrary values when argparse output is none.
-            # Host name
-            self._hostname = OsUtil.subproc_bash("hostname")[0].strip()
-            self._logger.warning(f"Hostname not passed in args, using the current hostname: '{self._hostname}'")
-            # User ID
-            try:
-                self._os_user_id = pwd.getpwuid(os.getuid()).pw_name
-            except KeyError as e:
-                self._logger.warning(f"Cannot get user ID from OS. Error: {str(e)}\nSetting an arbitrary user ID 'user_set_by_suco'.")
-        else:
-            if args_in.hostname:
-                self._hostname = args_in.hostname
-
-            if args_in.user_id:
-                self._os_user_id = args_in.user_id
-
-        # Python security https://docs.python.org/3.10/library/subprocess.html#popen-constructor
-        # for those executables that are (hopefully) available on any shell independent from the type of OS.
-
-        # TODO This member var, purpose of which particularly, is fairly undefined.
-        # Better way to manage un/found execs is wanted.
-        self._execs_found = self.docker_available(self._args_in)
-        self._setup_git()
-
-    def docker_available(self, args_in: argparse.Namespace):
-        """
-        @deprecated: `get_paths_execs` is planned to be deprecated throughout the entire package
-            as `subprocess` should be able to resolve just like the shell environment does, as long as
-            the values of `PATH` env var is properly passed.
-        """
-        self._logger.warning(f"'get_paths_execs' in ShellCapableOsSetup: '{args_in.skip_setup_docker=}'")
-        if not args_in.skip_setup_docker:
-            # Only when 'skip_setup_docker' is True.
-            # self.setup_docker(userid_os=self._os_user_id, skip=args_in.skip_setup_docker)
-            self._which_docker = OsUtil.which("docker")
-            self._logger.info(f"Path to 'docker' executable: {self._which_docker}")
-
-    def _setup_git(self):
-        """
-        @description:
-            Upon implementation, 'self._which_git' must be filled in with the concrete path of the executable of 'git'.
-        """
-        raise NotImplementedError()
-
-    def setup_vscode(self, path_installer: str):
-        raise NotImplementedError()
-
-    def swap_file(self, src_file: str, dest_file: str, suffix_backup=".org"):
-        """
-        @description: Swap 'dest_file' with 'src_file', which can be a symlink. Backup of 'dest_file' will be made alongside the swapped file.
-        @param src_file: Absolute path of the source file.
-        @param dest_file: Absolute path of the destination file.
-        """
-        raise NotImplementedError()
-
-    def setup_file(self, file_dispatch: ConfigDispach, overwrite=False):
-        try:
-            OsUtil.copy_a_file(
-                file_dispatch.path_source, file_dispatch.path_dest, is_symlink=file_dispatch.is_symlink, overwrite=overwrite)
-        except FileExistsError as e:
-            self._logger.warning("Target already exists. Moving on. \n{}".format(str(e)))
-        except FileNotFoundError as e:
-            raise
-
-    def setup_terminal_configs(self, abspath_local_perm_conf: str):
-        raise NotImplementedError("Terminal config setup needs to be implemented in the derived class.")
-
-    def setup_git_config(self, path_local_perm_conf):
-        path_user_home = pathlib.Path.home()
-
-        conf_gitconf = ConfigDispach(
-            path_source=os.path.join(path_local_perm_conf, "dot_gitconfig"),
-            path_dest=os.path.join(path_user_home, ".gitconfig",),
-            is_symlink=True)
-        conf_gitignore = ConfigDispach(
-            path_source=os.path.join(path_local_perm_conf, "dot_gitignore_global"),
-            path_dest=os.path.join(path_user_home, ".gitignore_global"),
-            is_symlink=True)
-        self.setup_file(conf_gitconf)
-        self.setup_file(conf_gitignore)
-
-    def setup_dropbox(self):
-        output, error, bash_return_code = OsUtil.subproc_bash("dropbox")
-        if bash_return_code == 0:
-            self._logger.info("Skipping Dropbox setup as it's already set up.")
-
-        FILENAME_DEB_DROPBOX = "dropbox_2022.12.05_amd64.deb"
-        url_deb = "https://linux.dropbox.com/packages/ubuntu/{}".format(FILENAME_DEB_DROPBOX)
-        os.chdir("/tmp")
-        OsUtil.subproc_bash("wget {}".format(url_deb))
-        cmd_install = "dpkg -i download?dl=packages%2Fubuntu%2F{}".format(FILENAME_DEB_DROPBOX)
-        OsUtil.subproc_bash(cmd_install, does_sudo=True)
-
-    def clone(self, repo_to_clone, dir_cloned_at, branch=""):
-        """
-        @return: Absolute path of the successfully cloned local repo.
-        @raise ValueError when some input is null
-        """
-        if not dir_cloned_at:
-            raise ValueError(f"Var 'dir_cloned_at' cannot be null.")
-        
-        _abs_path_local = os.path.join(dir_cloned_at, OsUtil.get_repo_basename_from_url(repo_to_clone))
-        if os.path.exists(_abs_path_local):
-            self._logger.warning(f"Skppig to git clone '{repo_to_clone}' as a local path '{_abs_path_local}' already exists." \
-                                 f"NOTE: This can result in the code of '{repo_to_clone}' may not get updated since the first time it was cloned" \
-                                 f"      To avoid that, you may want to manually delete '{_abs_path_local}'")
-            return _abs_path_local
-
-        self._logger.info(f"Cloning '{repo_to_clone}' into a local dir: '{dir_cloned_at}' so the abs local path will be '{_abs_path_local}.")
-        self.git_clone_impl(repo_to_clone, dir_cloned_at, branch)
-
-        # Verifying if perm conf repo is successfully cloned on the host, by checking to see if the path exists.
-        if not os.path.exists(dir_cloned_at):
-            raise FileNotFoundError(
-                f"At '{dir_cloned_at}', a local repo '{repo_to_clone}' is expected to be present in order to continue.")
-        return _abs_path_local
-
-    def _is_docker_setup(self):
-        bash_return_code = -1
-        _MSG_ERR = "Docker setup is not done yet"
-        try:
-            output, error, bash_return_code = OsUtil.subproc_bash(f"docker images")
-        except AttributeError as e:
-            raise RuntimeWarning(_MSG_ERR)
-        if bash_return_code == 0:
-            self._logger.info("Docker setup skipped as it's already set up.")
-        else:
-            raise RuntimeWarning("Docker setup is not done yet")
-        return bash_return_code
-
-    def _import_git(self):
-        """
-        @summary: Very adhoc method
-          In case `python3-git` module wasn't installed when this program started (so that `import git` failed when this file was read in),
-          re-importing Python's `git` module and let Python interpreter recognize the module to be loaded.
-          Ref. https://stackoverflow.com/a/19179497/577001
-        """
-        try:
-            importlib.reload(git)
-        except NameError as e:
-            globals()["git"] = importlib.import_module("git")
-            self.add_runtime_issue(e)
-
-    def _git_clone_py(self, repo_to_clone, dir_cloned_at):
-        git.Repo.clone_from(repo_to_clone, dir_cloned_at)
-
-    def git_clone_impl(self, repo_to_clone, dir_cloned_at):
-        raise NotImplementedError()
-
-    def setup_docker(self, userid_os, skip=False):
-        """
-        @param skip: Set 'True' when docker is not necessary e.g. running already inside a docker container.
-        """
-        raise NotImplementedError()
-
-    def setup_configs(self, host_config: HostConf, abs_path_confdir: str):
-        raise NotImplementedError()
-
-    def generate_symlinks(self, rootpath_symlinks, path_user_home):
-        raise NotImplementedError()
-
-    def common_symlinks(self, pairs_symlinks: list[ConfigDispach]):
-        for pair in pairs_symlinks:
-            self._logger.info(f"ConfigDispach: {pair}, pairs_symlinks: {pairs_symlinks}")
-            try:
-                self.setup_file(pair)
-            except FileNotFoundError as e:
-                _error_msg = f"""Source of symlink '{pair.path_source}' it not (yet) found on the local file system. 
-This is most notably ammendable by setting up local client executables of Dropbox and/or Google Drive."""
-                if pair.hint_enable:
-                    _error_msg += f"\nHint: {pair.hint_enable}"
-                self.add_runtime_issue(e)
-                self._logger.error(_error_msg)
-                if not pair.necessary:
-                    continue
-                raise e
-
-    def set_os_user_conf(
-            self,
-            path_local_conf_repo,
-            list_conf_files,
-            path_user_home_dir,
-            path_section_conf_dir):
-        """
-        @deprecated: Use 'OsUtil.copy_a_file' instead.
-        @summary: Setup configuration files under a Linux user's home directory. This method
-            should be capable of handling:
-            - Copying a file from dir 'a' to 'b'.
-            - Creating a config folder, if it does not exist.
-            - Creating a symlink if desired, with a custom name if specified.
-           Source path | Target path | symlink target (if different from the file/directory name)
-        @param path_user_home_dir: E.g. '/home/foo'
-        @param path_section_conf_dir: Secton of the path for the config directory under 'path_user_home_dir' E.g. '.config/autostart'
-        """
-        path_config_dir = os.path.join(path_user_home_dir, path_section_conf_dir)
-        if not os.path.exists(path_config_dir):
-            os.mkdir(path_config_dir)
-        for conf_file in list_conf_files:
-            path_file = os.path.join(path_local_conf_repo, conf_file)
-            shutil.copyfile(path_file, os.path.join(path_config_dir, conf_file))
-
-    def setup_ros_installer_src(self):
-        raise NotImplementedError()
-
-    def setup_rosdep_and_run(self, path_ws, pkg_rosdep="python3-rosdep", init_rosdep=False):
-        raise NotImplementedError()
-
-    def install_deps_adhoc(self, deb_pkgs=[], pip_pkgs=[], allow_pip_break=False, snap_pkgs: list[str]=[]):
-        """
-        @param allow_pip_break: If True, pip runs with '--break-system-packages' option.
-        """
-        raise NotImplementedError()
-
-    def setup_ssh(self, skip=True, path_local_conf_repo=""):
-        """
-        @param skip: If True, skip setting up SSH.
-        @param path_local_conf_repo: Path to the local configuration repo.
-        """
-        if skip:
-            self._logger.info("Skipping SSH setup as 'skip' is set to True.")
-            return
-
-        # Implement SSH setup logic here
-        raise NotImplementedError("SSH setup is not implemented yet.")
-
-    def nonrosdep_deps(self) -> tuple[list[str], list[str]]:
-        """
-        @return: Tuple of two lists: (list of deb packages, list of pip packages)
-        """
-        raise NotImplementedError()
-
-    def run(self, args, host_config, conf_repo_remote, conf_base_path):
-        """
-        @type args: (argparse' output)
-        @type host_cfg: HostConf
-        @param conf_repo: Absolute path URL of the repo to clone that contains host config.
-        @param conf_base_path: Path to a local location conf_repo to be cloned to.
-          Default is defined in each OS type class by "_PATH_BASE_CONF" variable.
-        """
-        self._logger.info("Update the host name as '{}'".format(self._hostname))
-
-        # Extract repo base name (e.g. 'xyz' from https://github.org/orgorg/xyz.git)
-        _repo_basename = OsUtil.get_repo_basename_from_url(conf_repo_remote)
-        _abs_path_repo_cloned_into = os.path.join(conf_base_path, _repo_basename)
-        _conf_repo_version = self._args_in.conf_repo_version
-        self._logger.debug(f"_abs_path_repo_cloned_into: {_abs_path_repo_cloned_into}")
-        self.clone(conf_repo_remote, _abs_path_repo_cloned_into, branch=_conf_repo_version)
-
-        if not self._args_in.skip_setup_docker:
-            self._logger.warning(f"{self._args_in.skip_setup_docker=}. Setting up Docker with user ID '{self._os_user_id}'.")
-            try:
-                self.setup_docker(userid_os=self._os_user_id, skip=self._args_in.skip_setup_docker)
-            except AttributeError as e:
-                _MSG_E = f"'setup_docker' method is incomp;lete. Moving on despite the error: {str(e)}"
-                self.add_runtime_issue(_MSG_E)
-
-        else:
-            self._logger.info(f"Skipping Docker setup as 'skip_setup_docker' is set to True (verify -> {self._args_in.skip_setup_docker}).")
-
-        try:
-            self.update_hostname(self._hostname)
-        except NotImplementedError as e:
-            self._logger.warning("{}\nIgnore and moving on for now.".format(str(e)))
-            self.add_runtime_issue(e)
-
-        self._user_home_dir = pwd.getpwuid(os.getuid()).pw_dir
-        self._logger.info(f"""Set home dir of the user at {self._user_home_dir} that will be the main user account on this computer.
-            For now the user account that is used to execute this process will be the main account.""")
-
-        # Installation by batch based on the list defined in package.xml.
-        self.setup_rosdep_and_run(args.path_local_conf_repo, init_rosdep=True)
-        # Install dependency that is not available via rosdep
-        _deps, _deps_pip = self.nonrosdep_deps()
-        try:
-            # TODO In DebianSetup, 'install_deps_adhoc' is already called within 'setup_rosdep_and_run',
-            # so this call here might be redundant with no good reason. This needs to be re-think-ed.
-            self.install_deps_adhoc(_deps, _deps_pip)
-        except RuntimeWarning as e:
-            self.add_runtime_issue(e)            
-        except RuntimeError as e:
-            self.add_runtime_issue(e)            
-
-        # This must be implemented for all OSes as the end result is crucial to my computer usage,
-        # therefore do NOT catch `NotImplementedError`.
-        try:
-            self.swap_caps_ctrl()
-        except RuntimeError as e:
-            self.add_runtime_issue(e)
-
-        _abs_path_confdir = os.path.join(args.path_local_conf_repo, args.path_conf_dir)
-        self._logger.debug(f"Abs_path_confdir: '{_abs_path_confdir}")
-
-        self.setup_terminal_configs(_abs_path_confdir)
-
-        try:
-            self.setup_ssh()
-        except RuntimeWarning as e:
-            self.add_runtime_issue(e)            
-        except Exception as e:
-            self.add_runtime_issue(e)
-
-        self.setup_git_config(path_local_perm_conf=_abs_path_confdir)
-
-        # Skip Google Chrome specific setting as it might come bundled already on Ubuntu.
-
-        # Skip synergy setting.
-
-        self.setup_dropbox()
-
-        self.create_data_dir(
-            [os.path.join(self._user_home_dir, self._DIR_DROXBOX_CONTAINER),
-             os.path.join(self._user_home_dir, args.path_symlinks_dir)])
-        _pairs_symlinks = self.generate_symlinks(
-            rootpath_symlinks=os.path.join(self._user_home_dir, args.path_symlinks_dir),
-            path_user_home=self._user_home_dir)
-        self.common_symlinks(_pairs_symlinks)
-
-        self.setup_configs(host_config, abs_path_confdir=_abs_path_confdir)
-
-        _msg_endroll = args.msg_endroll if args.msg_endroll else "Setup finished."
-        self._logger.info(_msg_endroll)
-
-        self.setup_vscode(path_installer=args.path_vscode_installer)
-
-
-class DebianSetup(ShellCapableOsSetup):
-    _APTPKG_ROSDEP2 = "python3-rosdep2"
-    _DEB_CAPS_CTRL_UTIL = "gnome-tweaks"
-    _DEBS_MOZC = ["emacs-mozc", "emacs-mozc-bin", "ibus-mozc", "mozc-utils-gui", "mozc-server"]
-    _DEBIAN_DEB_DEPS = [
-                "aptitude",
-                "colorized-logs",
-                "dconf-editor",
-                "evince",
-                "flameshot",
-                _DEB_CAPS_CTRL_UTIL,  # Primarily for swapping Caps and Ctrl keys
-                "ibus",
-                "libavahi-compat-libdnssd1",
-                "pdftk-java",
-                "pidgin",
-                "psensor",
-                #"python3-rosdep",  # Without ROS' apt source, apt would install python3-rosdep2, which is NOT the officially maintained pkg. See https://discourse.ros.org/t/upstream-packages-increasingly-becoming-a-problem/10902/25
-                "ptex-base",
-                "synaptic",
-                "xbindkeys",
-                "xsel",     # https://github.com/kinu-garage/hut_10sqft/issues/1077
-                "whois",
-                ] + _DEBS_MOZC
-    _OS_TYPE = OsUtil.TYPE_LINUX_DISTRO_DEBIAN
-    _PIP_PKGS = ["pipx"]
-
-    def __init__(self, os_name=_OS_TYPE, args_in: argparse.Namespace=None):
-        # This variable might be accessed in `super().__init__` so
-        # defined prior to the call. Might not be a good practice though.
-        self._apt_updated = False
-
-        super().__init__(os_name, args_in)
-
-    @property
-    def apt_updated(self):
-        return self._apt_updated
-
-    @apt_updated.setter
-    def apt_updated(self, value):
-        self._apt_updated = value
-
-    def setup_vscode(self, path_installer: str) -> str:
-        """
-        @return: Path to the 'code' executable, None if installation was unsuccessful.
-        @see https://code.visualstudio.com/blogs/2020/12/03/chromebook-get-started
-        @raise ReferenceError: if setting up failed.
-        """
-        if not path_installer:
-            raise ValueError(f"'path_installer' is empty.")
-
-        self.install_deps_adhoc("gnome-keyring")
-        cmd_install = f"dpkg -i {path_installer}"
-        OsUtil.subproc_bash(cmd_install, does_sudo=True)
-
-        # Returning the verification result
-        return OsUtil.which("code")
-
-    def setup_terminal_configs(self, abspath_local_perm_conf: str):
-        _CONFFILE_NAME_SHORTCUT = "terminal_shortcuts.dconf"
-        _abspath_conf_dir = os.path.join(abspath_local_perm_conf, "dconf")
-        _conf_abspath = os.path.join(_abspath_conf_dir, _CONFFILE_NAME_SHORTCUT)
-        _cmd = f"dconf load /org/gnome/terminal/ < {_conf_abspath}"
-        OsUtil.subproc_bash(_cmd, does_sudo=False, print_stdout_err=True, logger=self._logger)
-
-        _MSG_NOTE_TERMINAL_CONF_VISUAL_NOT_DONE = (f"""Configs of the visual of the Terminal needs to be done manually """
-                                                   """ either using (recommended) Terminal's GUI on 'Preference' or using `dconf` and the premade config files,"""
-                                                   f""" which you can find in '{_abspath_conf_dir}'. See https://github.com/kinu-garage/hut_10sqft/issues/174#issuecomment-3003542573""")
-        self._logger.warning(_MSG_NOTE_TERMINAL_CONF_VISUAL_NOT_DONE)
-        self.add_runtime_issue(_MSG_NOTE_TERMINAL_CONF_VISUAL_NOT_DONE)
-
-    def setup_ros_installer_src(self):
-        self._logger.warning(f"On '{self._OS_TYPE}' no prebuilt ROS installer pkgs are available so skipping.")
-
-    def exec_rosdep_update(self, path_ws, pkg_rosdep=_APTPKG_ROSDEP2, init_rosdep=False):
-        """
-        @summary: As of 202505 this method is only targetting Debian/Ubuntu OSes.
-        @param init_rosdep: If `True`, then `rosdep init` also executes.
-        """
-        self.setup_ros_installer_src()
-
-        if init_rosdep:
-            OsUtil.setup_rosdep()
-        os.chdir(path_ws)
-        self._logger.info(f"Changed directory to '{path_ws}' to run 'rosdep install' against the manifest that defines dependencies")
-        output, error, bash_return_code = OsUtil.subproc_bash("rosdep install --from-paths . --ignore-src -r -y")
-        if bash_return_code != 0:
-            self.add_runtime_issue(f"'rosdep install' failed.\n\tOutput: {output}\n\tError: {error}")
-        else:
-            self.add_runtime_issue(f"'rosdep install' succeeded.\n\tOutput: {output}\n\tError: {error}")
-        
-    def setup_rosdep_and_run(self, path_ws, pkg_rosdep=_APTPKG_ROSDEP2, init_rosdep=False):
-        """
-        @note: For Debian OS, no official prebuilt rosdep installer via apt is available,
-          but a community version 'python3-rosdep2' maintained by a long-term community member (Jochen S.) is avaialble
-          so using it for now. But for Ubuntu 'python3-rosdep' (without 2 at the end) is the official and should be used.
-        """
-        self.setup_ros_installer_src()
-        # Install deb dependencies that cannot be installed in the batch
-        # installation step that is planned later in this sequence.
-        self.install_deps_adhoc(deb_pkgs=["python3-pip", pkg_rosdep], pip_pkgs=self._PIP_PKGS)
-
-        self.exec_rosdep_update(path_ws, pkg_rosdep, init_rosdep)
-
-    def nonrosdep_deps(self) -> tuple[list[str], list[str]]:
-        """@override"""
-        return self._DEBIAN_DEB_DEPS, self._PIP_PKGS
-
-    def _install_deps_adhoc_debian(self, deb_pkgs=[], pip_pkgs=[], allow_pip_break=False):
-        OsUtil.apt_install(deb_pkgs, self._logger)
-        self._logger.info(f"pip_pkgs: {pip_pkgs}")
-        OsUtil.install_pip_adhoc(pip_pkgs, allow_break=allow_pip_break)
-        # TODO self.add_runtime_issue(f"'rosdep install' failed.\n\tOutput: {output}\n\tError: {error}")
-
-    def install_deps_adhoc(self, deb_pkgs, pip_pkgs="", allow_pip_break=False, snap_pkgs: list[str]=[]):
-        """
-        @summary: Install the packages that cannot be installed by batch using
-            'rosdep install'. Example is 'python3-rosdep' itself.
-        @param deb_pkgs: [str] intended but if a str without being in a list format nor whitespace,
-            then it should be accepted as well, as it'll be internally converted to a list.
-        @param pip_pkgs: Same applies as 'deb_pkgs'.
-        @param allow_break: If True, `pip` runs with '--break-system-packages' option.
-        """
-        self.apt_update()
-        self._install_deps_adhoc_debian(deb_pkgs, pip_pkgs, allow_pip_break)
-
-    def create_data_dir(self, dirs_tobe_made):
-        self._logger.info("Making directories historically been in use: {}".format(dirs_tobe_made))
-        for dir in dirs_tobe_made:
-            try:
-                os.mkdir(dir)
-            except FileExistsError as e:
-                self._logger.warning("{}\nIgnore and moving on for now.".format(str(e)))
-                self.add_runtime_issue(e)
-
-    def git_clone_impl(self, repo_to_clone, dir_cloned_at, branch=""):
-        _option = ""
-        if branch:
-            _option = "-b" + " " + branch
-        OsUtil.subproc_bash(f"git clone {repo_to_clone} {dir_cloned_at} {_option}", does_sudo=False, print_stdout_err=True)
-
-    def setup_ssh(self, skip=False, path_local_conf_repo=""):
-        self._logger.info(f"On {OsUtil.TYPE_OS_LINUX} type of OS, where 'rosdep install' should function (?), SSH server should be enabled when a set of relevant pkgs get installed via 'rosdep'.")
-
-    def setup_oracle_java(self):
-        self._logger.warning("""The following should be done manually, mainly due to license operation that is hard to automate, in order to set up Oracle Java that is required by Eclipse:
-
-    # Refs:
-    # - http://askubuntu.com/a/651045/24203
-    # - http://superuser.com/a/939651/106974
-    ## sudo add-apt-repository ppa:webupd8team/java
-    ## apt update && apt-get install -y oracle-java8-installer
-    ## sudo apt install oracle-java8-set-default
-""")
-
-    def run(self, args, host_config, conf_repo_remote, conf_base_path=""):
-        super().run(args, host_config, conf_repo_remote, conf_base_path)
-        self.setup_oracle_java()
-
-    def apt_update(self):
-        if self.apt_updated:
-            self._logger.warning("'apt update' was already done before. Skipping")
-            return
-        cmd = f"apt update"
-        _out, _err, retcode = OsUtil.subproc_bash(cmd, does_sudo=True)
-        if retcode != 0:
-            raise subprocess.CalledProcessError(
-                returncode = retcode,
-                cmd = cmd,
-                stderr = _err)
-        self.apt_updated = True
-
-    def setup_docker(self, userid_os, skip=False):
-        try:
-            if skip or self._is_docker_setup():
-                self._logger.info(f"Looks like Docker setup is already completed.")
-                return
-        except RuntimeWarning as e:
-            self._logger.warning(f"Issue found in setting up Docker but continuing to do so. Source of the error: {str(e)}")
-            self.add_runtime_issue(e)
-        if not self._execs_found:
-            self.add_runtime_issue("Not all necessary executables is found. Aborting setting up Docker.")
-            return
-
-        OsUtil.subproc_bash("groupadd docker", does_sudo=True)
-        OsUtil.subproc_bash("usermod -aG docker {}".format(userid_os), does_sudo=True)
-
-        # From https://docs.docker.com/engine/installation/linux/ubuntulinux/
-        OsUtil.subproc_bash(f"apt-key adv --keyserver hkp://p80.pool.sks-keyservers.net:80 --recv-keys 58118E89F3A912897C070ADBF76221572C52609D", does_sudo=True)
-        OsUtil.subproc_bash(f'echo "deb https://apt.dockerproject.org/repo ubuntu-`lsb_release -sc` main" > /etc/apt/sources.list.d/docker.list', does_sudo=True)
-        self.apt_update()
-        OsUtil.subproc_bash(f"apt purge lxc-docker", does_sudo=True)
-        OsUtil.subproc_bash(f"apt-cache policy docker-engine")
-        OsUtil.subproc_bash(f"apt install linux-image-extra-$(uname -r)", does_sudo=True)
-        # Workaround found at http://stackoverflow.com/questions/22957939/how-to-answer-an-apt-get-configuration-change-prompt-on-travis-ci-in-this-case
-        OsUtil.subproc_bash(f'apt -q -y -o Dpkg::Options::="--force-confdef" -o Dpkg::Options::="--force-confnew" install docker-engine',
-                            does_sudo=True, non_interactive=True)
-        OsUtil.subproc_bash(f"service docker start", does_sudo=True)
-        OsUtil.subproc_bash(f"unset $DEBIAN_FRONTEND")
-        OsUtil.subproc_bash(f'docker run hello-world && echo "docker seems to be installed successfully." || (echo "Something went wrong with docker installation."; RESULT=1', does_sudo=True)
-    
-    def _setup_git(self):
-        self.install_deps_adhoc(deb_pkgs=["python3-git"])
-        # If git had not been installed yet prior to the one line above,
-        # then its executable hadn't been available either.
-        self._which_git = OsUtil.which("git")
-
-    def setup_configs(self, host_config: HostConf, abs_path_confdir: str):
-        pairs_conf_autostart = [
-            ConfigDispach(
-                path_source=os.path.join(abs_path_confdir, "gnome-system-monitor.desktop"),
-                path_dest=os.path.join(self._user_home_dir, ".gconf/apps"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(abs_path_confdir, "indicator-multiload.desktop"),
-                path_dest=os.path.join(self._user_home_dir, ".config", "autostart"),
-                is_symlink=True),
-            ]
-        for conf in pairs_conf_autostart:
-            self.setup_file(conf)
-
-        pairs_conf_bash = [
-            ConfigDispach(
-                path_source=os.path.join(abs_path_confdir, "bash", host_config.bash_cfg),
-                path_dest=os.path.join(self._user_home_dir, ".bashrc"),
-                is_symlink=True),
-            ]
-        for c in pairs_conf_bash:
-            self.setup_file(c, overwrite=True)
-
-        pairs_conf_tools = [
-            ConfigDispach(
-                path_source=os.path.join(abs_path_confdir, "gnome-system-monitor.desktop"),
-                path_dest=os.path.join(self._user_home_dir, ".gconf/apps"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(abs_path_confdir, "indicator-multiload.desktop"),
-                path_dest=os.path.join(self._user_home_dir, ".gconf/apps"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(abs_path_confdir, "tmux_default.conf"),
-                path_dest=os.path.join(self._user_home_dir, ".tmux.conf"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(abs_path_confdir, "emacs", host_config.emacs_cfg),
-                path_dest=os.path.join(self._user_home_dir, ".emacs"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(abs_path_confdir, "dot_xbindkeysrc"),
-                path_dest=os.path.join(self._user_home_dir, ".xbindkeysrc"),
-                is_symlink=True),
-            ]
-        for c in pairs_conf_tools:
-            self.setup_file(c)
-
-    def verify_deb_installed(self, deb_pkg_name: str):
-        """
-        @summary: Verifies if `deb_pkg_name` package is installed.
-        @exception RuntimeError: If `deb_pkg_name` package is not installed.
-        """
-        cmd = f"apt-cache policy {deb_pkg_name}"
-        output, error, bash_return_code = OsUtil.subproc_bash(cmd, does_sudo=True)
-        if bash_return_code != 0:
-            raise RuntimeError(f"Package '{deb_pkg_name}' is not installed.")
-        self._logger.info(f"Package '{deb_pkg_name}' seems already installed. \n\tCMD executed: {cmd}\n\tOutput: {output}")
-
-    def swap_caps_ctrl(self):
-        """
-        @summary: Installs S/Ws that are needed to swap Caps Lock and Ctrl keys, BUT configuring it needs to be done manually.
-        """
-        self.verify_deb_installed(self._DEB_CAPS_CTRL_UTIL)
-        _URL_INSTRUCTION_CAPS_CTRL = "https://github.com/kinu-garage/hut_10sqft/issues/1230#issuecomment-2994825273"
-        self._logger.info(f"Following {_URL_INSTRUCTION_CAPS_CTRL}, setup manually the swap of Caps Lock and Ctrl keys.")
-
-
-class ChromeOsSetup(DebianSetup):
-    _DIRNAME_GDRIVE = "GoogleDrive"
-    _DIRNAME_LOCAL_DIR = "MyFiles"
-    _DIRNAME_LOCAL_DOWNLOADS = "Downloads"
-    _HINT_ENABLE_MOUNT_GENERIC = "Likely any of '{}' path is not yet mounted on the Linux container. To mount,\n" \
-        "1. Open 'File app' on the host ChromeOS.\n" \
-        "2. on the left pane (list of directories) expand {}." \
-        "3. On any top-level folder you'd like to mount on to your Linux container, right-click then choose 'Manage Linux Sharing' then the app should show a confirmation msg.\n" \
-        "4. Verify on terminal on a Linux container that the directory is found by running 'ls -l /mnt/chromeos/{}'.\n"
-    _HINT_ENABLE_MOUNT_GDRIVE = _HINT_ENABLE_MOUNT_GENERIC.format(_DIRNAME_GDRIVE, _DIRNAME_GDRIVE, _DIRNAME_GDRIVE)
-    _HINT_ENABLE_MOUNT_LOCAL_DOWNLOADS = _HINT_ENABLE_MOUNT_GENERIC.format(_DIRNAME_LOCAL_DOWNLOADS, _DIRNAME_LOCAL_DOWNLOADS, os.path.join(_DIRNAME_LOCAL_DIR, _DIRNAME_LOCAL_DOWNLOADS))
-    _OS_TYPE = OsUtil.TYPE_OS_CHROMEOS
-
-    def __init__(self, os_name=_OS_TYPE, args_in: argparse.Namespace=None):
-        super().__init__(os_name, args_in)
-
-    def setup_dropbox(self):
-        self._logger.warning(
-            f"Skipping Dropbox setup on {self._OS_TYPE}, as it runs on the Chrome OS host without allowing to mount the directory onto Linux mode.")
-
-    def generate_symlinks(self, rootpath_symlinks: str, path_user_home=""):
-        """
-        @param rootpath_symlinks: Path to the directory that is designed to host the list of symlinks e.g. '~/link'.
-        """
-        pairs_symlinks = [
-            ConfigDispach(
-                path_source=os.path.join(os.path.sep, "mnt" ,"chromeos", self._DIRNAME_GDRIVE, "MyDrive"),
-                path_dest=os.path.join(rootpath_symlinks, self._DIRNAME_GDRIVE), is_symlink=True, necessary=True, hint_enable=self._HINT_ENABLE_MOUNT_GDRIVE),
-            ConfigDispach(
-                path_source=os.path.join(rootpath_symlinks, self._DIRNAME_GDRIVE, "30y-130s"),
-                path_dest=os.path.join(rootpath_symlinks, "30y-130s"), is_symlink=True, necessary=True, hint_enable=self._HINT_ENABLE_MOUNT_GDRIVE),
-            ConfigDispach(
-                path_source=os.path.join(rootpath_symlinks, self._DIRNAME_GDRIVE, "Current"),
-                path_dest=os.path.join(rootpath_symlinks, "Current"), is_symlink=True, necessary=True, hint_enable=self._HINT_ENABLE_MOUNT_GDRIVE),
-            ConfigDispach(
-                path_source=os.path.join(rootpath_symlinks, self._DIRNAME_GDRIVE, "Career", "academicDoc"),
-                path_dest=os.path.join(rootpath_symlinks, "academicDoc"), is_symlink=True, necessary=True, hint_enable=self._HINT_ENABLE_MOUNT_GDRIVE),
-            ConfigDispach(
-                path_source=os.path.join(rootpath_symlinks, self._DIRNAME_GDRIVE, "Career", "MOOC"),
-                path_dest=os.path.join(rootpath_symlinks, "MOOC"), is_symlink=True, necessary=True, hint_enable=self._HINT_ENABLE_MOUNT_GDRIVE),
-            ConfigDispach(
-                path_source=os.path.join(os.path.sep, "mnt" ,"chromeos", self._DIRNAME_LOCAL_DIR, self._DIRNAME_LOCAL_DOWNLOADS),
-                path_dest=os.path.join(rootpath_symlinks, "chrome-host_downloads"), is_symlink=True, necessary=True, hint_enable=self._HINT_ENABLE_MOUNT_LOCAL_DOWNLOADS),
-            ]
-        self._logger.debug(f"pairs_symlinks: type: {type(pairs_symlinks)}, content: {pairs_symlinks}")
-        return pairs_symlinks
-
-
-class UbuntuOsSetup (DebianSetup):
-    _OS_TYPE = OsUtil.TYPE_LINUX_DISTRO_UBUNTU
-    _UBUNTU_DEB_DEPS = [
-        "gnome-screenshots",        
-        "googleearth-package",
-        "gtk-recordmydesktop",
-        "ibus-el",
-        "indicator-multiload",
-        "peek",
-        "python-software-properties",  # From http://askubuntu.com/a/55960/24203 primarilly for Oracle Java for Eclipse
-        "ptex-bin",
-        "sysinfo",        
-    ]    
-    _EXTERNAL_STORAGE_KUDU1 = "Evo840SSD"
-    _PKGS_SNAP = ["docker", "yt-dlp"]  # TODO Needs a better way specify this list of pkgs.
-
-    def __init__(self, os_name=_OS_TYPE, args_in: argparse.Namespace=None):
-        super().__init__(os_name, args_in)
-        self.ubuntu_desktop_cleanup()
-
-    def install_deps_adhoc(self, deb_pkgs=[], pip_pkgs=[], allow_pip_break=False, snap_pkgs: list[str]=_PKGS_SNAP):
-        if not deb_pkgs:
-            deb_pkgs = self._DEBIAN_DEB_DEPS + self._UBUNTU_DEB_DEPS
-        self._install_deps_adhoc_debian(deb_pkgs, pip_pkgs, allow_pip_break)
-
-        # Take care of `snap` packages
-        snap_pkgs_failed = []
-        for snap_pkg in snap_pkgs:
-            try:
-                self.setup_snap_pkgs(snap_pkg)
-            except RuntimeError as e:
-                snap_pkgs_failed.append(snap_pkg)
-        if snap_pkgs_failed:
-            self.add_runtime_issue(f"The following `snap` pkgs failed to install: {snap_pkgs_failed}.")
-
-    def ubuntu_desktop_cleanup(self):
-        dirs_tobe_removed = ["Documents", "Music", "Pictures", "Public", "Templates", "Videos"]
-        self._logger.warning("Deleting Ubuntu's default directories: {}".format(dirs_tobe_removed))
-        for dir in dirs_tobe_removed:
-            try:
-                shutil.rmtree(dir)
-            except FileNotFoundError as e:
-                self._logger.warning("File/Dir '{}' does not exist. Moving on without deleting it.".format(dir))
-                self.add_runtime_issue(e)
-
-    def generate_symlinks(self, rootpath_symlinks, path_user_home):
-        pairs_symlinks = [
-            ConfigDispach(  # Many symlinks depend on this symlink.
-                path_source=os.path.join(path_user_home, self._DIR_DROXBOX_CONTAINER, "Dropbox", "GoogleDrive"),
-                path_dest=os.path.join(rootpath_symlinks, "GoogleDrive"),
-                is_symlink=True),
-            ConfigDispach(  # Some others depend on this symlink.
-                path_source=os.path.join(path_user_home, "link", "GoogleDrive", "30y-130s"),
-                path_dest=os.path.join(rootpath_symlinks, "30y-130s"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(path_user_home, self._DIR_DROXBOX_CONTAINER, "Dropbox", "pg", "myDevelopment", "git_repo"),
-                path_dest=os.path.join(rootpath_symlinks, "git_repos"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(path_user_home, "link", "GoogleDrive", "Career"),
-                path_dest=os.path.join(rootpath_symlinks, "Career"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(path_user_home, "link", "GoogleDrive", "Current"),
-                path_dest=os.path.join(rootpath_symlinks, "Current"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(path_user_home, "link", "Career", "engineering", "ARIAC"),
-                path_dest=os.path.join(rootpath_symlinks, "ARIAC"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(path_user_home, "link", "Career", "MOOC"),
-                path_dest=os.path.join(rootpath_symlinks, "MOOC"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(path_user_home, "link", "Career", "academicDoc"),
-                path_dest=os.path.join(rootpath_symlinks, "academicDoc"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(path_user_home, "link", "git_repos", "ROS", "cws_base"),
-                path_dest=os.path.join(rootpath_symlinks, "ROS"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(path_user_home, "link", "30y-130s", "schools_children", "GJLS"),
-                path_dest=os.path.join(rootpath_symlinks, "GJLS"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(path_user_home, "link", "git_repos", "ROS", "cws_utakata"),
-                path_dest=os.path.join(rootpath_symlinks, "cws_utakata"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=(os.path.sep + os.path.join("media", self._os_user_id, self._EXTERNAL_STORAGE_KUDU1)),
-                path_dest=os.path.join(rootpath_symlinks, self._EXTERNAL_STORAGE_KUDU1),
-                necessary=False,
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(path_user_home, self._DIR_DROXBOX_CONTAINER, "Dropbox", "My Mac (tork-mac1)"),
-                path_dest=os.path.join(rootpath_symlinks, "dbox_mac1"),
-                is_symlink=True),
-            ]
-        return pairs_symlinks
-
-    def set_ros_apt_source(self, 
-                           path_aptsrc_file="/etc/apt/sources.list.d/ros2.list",
-                           path_os_release = "/etc/os-release",
-                           key_os_code = "VERSION_CODENAME="):
-        _URL_ROS2_UBUNTU_INSTALL_DEP = "https://docs.ros.org/en/jazzy/Installation/Ubuntu-Install-Debs.html"
-        self._logger.warning(f"At some point after ROS2 Foxy, the apt source setting for ROS2 on Ubuntu has changed. Follow manually the instruction at {_URL_ROS2_UBUNTU_INSTALL_DEP}. "
-                             f"If the ROS2 setup is not done, then the future steps that depend on ROS2 apt source setting might fail.")
-
-    def _set_ros_apt_source(self, 
-                           path_aptsrc_file="/etc/apt/sources.list.d/ros2.list",
-                           path_os_release = "/etc/os-release",
-                           key_os_code = "VERSION_CODENAME="):
-        """
-        @deprecated: This method uses `ros2.list` in apt source, which seems to be outdated in ROS2 newer than Foxy.
-        @summary Create
-        @param key_os_code: Likely must end with '=', at least so as of Ubuntu 22.04.
-        @exception IOError: When an error happens while writing to 'path_aptsrc_file'
-        """
-        _URL_ROS2_UBUNTU_INSTALL_DEP = "https://docs.ros.org/en/foxy/Installation/Ubuntu-Install-Debians.html#install-ros-2-packages"
-        _CMD_ROS2_UBUNTU_INSTALL_DEP = 'echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/ros-archive-keyring.gpg] http://packages.ros.org/ros2/ubuntu $(. /etc/os-release && echo $UBUNTU_CODENAME) main" | sudo tee /etc/apt/sources.list.d/ros2.list > /dev/null'
-        _os_codename = ""
-        # Get OS codename        
-        with open(path_os_release, "r") as file_os_code:
-            str_file = file_os_code.readlines()
-            for line in str_file:
-                self._logger.info(f"line= {line}")
-                if line.startswith(key_os_code):
-                    # E.g. From a string 'VERSION_CODENAME=jammy', the following line extracts and put 'jammy' in '_os_codename'.
-                    _os_codename = line[line.index(key_os_code) + len(key_os_code):]
-                    # Cleaning e.g. remove line feed at the end.
-                    _os_codename = _os_codename.strip()
-                    break
-        if not _os_codename:
-            raise RuntimeError(f"OS code name not found in the file '{path_os_release}'")
-        res_oscode, err, retcode = OsUtil.subproc_bash("dpkg --print-architecture")
-
-        try:
-            with open(path_aptsrc_file, "w") as file_apt_src:
-                file_apt_src.write(
-                    f"deb [arch={res_oscode} signed-by=/usr/share/keyrings/ros-archive-keyring.gpg] http://packages.ros.org/ros2/ubuntu {_os_codename} main")
-        except IOError as e:
-            raise e
-        except PermissionError as e:
-            raise RuntimeError(f"Permission error while writing to '{path_aptsrc_file}'. "
-                               f"If the file is NOT yet present there, make an empty file by `sudo touch {path_aptsrc_file}`, then"
-                               f"follow the instruction at {_URL_ROS2_UBUNTU_INSTALL_DEP} (the distro is Focal, which is EoL but after that ROS seems to have switched the apt setting on Ubuntu, "
-                               f"which needs to be reviewed first before adjusting to it). Error: {str(e)}")
-
-    def setup_ros_installer_src(self):
-        self.set_ros_apt_source()
-        cmd_obtain_apt_key_rosdep = f"wget http://packages.ros.org/ros.key"
-        cmd_set_apt_key_rosdep = f"apt-key add ros.key"
-        OsUtil.subproc_bash(cmd_obtain_apt_key_rosdep)
-        OsUtil.subproc_bash(cmd_set_apt_key_rosdep, does_sudo=True)
-        self.apt_update()
-
-    def setup_rosdep_and_run(self, path_ws, pkg_rosdep="python3-rosdep", init_rosdep=False):
-        if pkg_resources == self._APTPKG_ROSDEP2:
-            self._logger.warning(f"On Ubuntu, relying on '{self._APTPKG_ROSDEP2}', which is unofficially maintained, is not recommended. For now moving foward though.")
-        self.exec_rosdep_update(path_ws, pkg_rosdep, init_rosdep)
-
-    def setup_snap_pkgs(self, snap_pkg: str):
-        """
-        @summary: Install specific `snap` packages that are not available via apt and other package managers.
-        """
-        cmd = f"snap install {snap_pkg}"
-        output, error, bash_return_code = OsUtil.subproc_bash(cmd, does_sudo=True)
-        if bash_return_code != 0:
-            raise RuntimeError(f"Failed to install snap package '{snap_pkg}'.\n\tOutput: {output}\n\tError: {error}")
-        self._logger.info(f"Successfully installed snap package '{snap_pkg}'.\n\tOutput: {output}\n\tError: {error}")
-
-class MacOsSetup(AbstCompSetupFactory):
-    _OS_TYPE = OsUtil.TYPE_OS_MACOS
-    def __init__(self, os_name=_OS_TYPE, args_in: argparse.Namespace=None):
-        super().__init__(os_name)
-
-    def install_deps_adhoc(self, deb_pkgs=[], pip_pkgs=[], allow_pip_break=False, snap_pkgs: list[str]=[]):
-        raise RuntimeWarning("TBD On MacOS maybe set up brew first, then install the dependencies via brew, pip, etc.")
-
-    def generate_symlinks(self, rootpath_symlinks: str, path_user_home=""):
-        pairs_symlinks = [
-            ConfigDispach(  # Many symlinks depend on this symlink.
-                path_source=os.path.join(path_user_home, self._DIR_DROXBOX_CONTAINER, "Dropbox", "GoogleDrive"),
-                path_dest=os.path.join(rootpath_symlinks, "GoogleDrive"),
-                is_symlink=True),
-            ConfigDispach(  # Some others depend on this symlink.
-                path_source=os.path.join(path_user_home, "link", "GoogleDrive", "30y-130s"),
-                path_dest=os.path.join(rootpath_symlinks, "30y-130s"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(path_user_home, "link", "GoogleDrive", "Current"),
-                path_dest=os.path.join(rootpath_symlinks, "Current"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(path_user_home, "link", "GoogleDrive", "Career", "MOOC"),
-                path_dest=os.path.join(rootpath_symlinks, "MOOC"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(path_user_home, "link", "Career", "academicDoc"),
-                path_dest=os.path.join(rootpath_symlinks, "academicDoc"),
-                is_symlink=True),
-            ConfigDispach(
-                path_source=os.path.join(path_user_home, "link", "30y-130s", "schools_children", "GJLS"),
-                path_dest=os.path.join(rootpath_symlinks, "GJLS"),
-                is_symlink=True),
-            ]
-        return pairs_symlinks
-
-    def setup_ssh(self, skip=False, path_local_conf_repo=""):
-        _MSG_INSTRUCTION_ENABLE_SSH_SERVER = "On terminal run 'sudo systemsetup -setremotelogin on'. \
-            If you get an error that looks like:\n \
-            'Turning Remote Login on or off requires Full Disk Access privileges'\n\n \
-           then go to Settings > Security & Privacy > Privacy > Full Disk Access; \
-           then select Applications > Utilities > Terminal from the file picker². Then execute the command again, \
-           it should be in the history so you might just need to press arrow upwards."
-
-        raise RuntimeWarning(f"Enabling SSH server on MacOS might have to be done manually. See https://superuser.com/a/1764825/106974, \
-                             or the text copied from there below:\n{_MSG_INSTRUCTION_ENABLE_SSH_SERVER}")
-
-    def setup_vscode(self, path_installer: str):
-        raise RuntimeWarning("Skipping as no plan to use this host for the development.")
-
-
-class CompInitSetup():
-    """
-    @summary TBD
-    """
-    HOSTNAME_BRYA = "130s-brya"
-    HOSTNAME_C13_MORPH = "130s-C13-Morph"
-    HOSTNAME_MAC1 = "130s-mac1"
-    HOSTNAME_OPFYDE_RPI5 = "opfyde-rpi5"
-    HOSTNAME_P16S = "130s-p16s-2"
-    HOSTNAME_ZORK16 = "130s-zork16"
-
-    _LOGGER_NAME = "CompInitSetup-logger"
-    # Name of the local repo that stores the config and will have to be
-    # available for the entire life time of the OS. 
-    _REPO_PERMANENT_CONFIG = "hut_10sqft"
-    _FOLDER_CONF_PERM_REPO = "config"
-    _PATH_FOLDER_CONF = os.path.join(pathlib.Path.home(), "." + _FOLDER_CONF_PERM_REPO)
-    # Un-expanded version of this looks like '~/.config/hut_10sqft'
-    _PATH_DEFAULT_PERMANENT_CONF_REPO = os.path.join(_PATH_FOLDER_CONF, _REPO_PERMANENT_CONFIG)
-    # Un-expanded version of this looks like 'hut_10sqft/config'
-    _PATH_DEFAULT_CONFIG_CONFDIR = os.path.join(_REPO_PERMANENT_CONFIG, _FOLDER_CONF_PERM_REPO)
-    _PATH_SYMLINKS_DIR = "link"  # e.g. ~/link
-    # Messages for stdout
-    _MSG_CONSOLE_TOOL_INTRO = """This tool is for setting up a Linux-based personal computer.
-     It does the following: 1) Installs dependency (which must be defined in package.xml). 
-     2) Create symlinks to config files, which are provided in hut_10sqft local git
-       repo i.e. (Having Khut_10sqft somewhere on the host is required)."""
-    _MSG_PATH_PERMCONF_REPO = f"""Path to the FINAL location of '{_REPO_PERMANENT_CONFIG}' local repo.
- If not passed then the path will be the default {_PATH_DEFAULT_PERMANENT_CONF_REPO}, 
-which is for {DebianSetup._OS_TYPE}."""
-    _MSG_PATH_CONF_DIR = f"""Path to the the config folder within the '{_REPO_PERMANENT_CONFIG}' repo.
- If not passed then the path will be the default {_PATH_DEFAULT_CONFIG_CONFDIR}."""
-    _MSG_ARG_PATH_COMMON_SYMLINKS = f"""Path to the folder that contains symlinks.
- If not passed then the path will be the default {_PATH_SYMLINKS_DIR}."""    
-    _MSG_ARG_USERID = """User ID on the OS that will be mainly used. While this
-is optional, it is recommened to specify. If nothing passed, then the tool
-treats the user ID tha is used to execute this tool as the main user."""
-    _MSG_ARG_BASE_CONF_PATH = """Path where the conf repo will be cloned into.
- Modifying it is an advanced/bold move, and behavior with the modified path is not planned to be tested as of 2024/08."""
-    _URL_CONFREPO = f"https://github.com/kinu-garage/{_REPO_PERMANENT_CONFIG}.git"
-    _PATH_VSCODE_INSTALLER = pathlib.Path("~/link/GoogleDrive/lifeinfra/computer/installer/vscode/code_1.103.2-1755709794_amd64.deb").expanduser()
-
-    def __init__(self):
-        self._logger = logging.getLogger(self._LOGGER_NAME)
-        log_handler = logging.StreamHandler()
-        self._logger.setLevel(logging.DEBUG)  # Needs changed
-        self._logger.addHandler(log_handler)
-
-    def _cli_args(self):
-        """
-        @rtype: argparse.Namespace
-        """
-        parser = argparse.ArgumentParser(description=self._MSG_CONSOLE_TOOL_INTRO)
-        # Optional but close to required args
-        parser.add_argument("--hostname", required=True, help="Specify in case you need to modify the host name.")
-        parser.add_argument("--msg_endroll", help="Specify the message string that will be printed at the end in case of need.")
-        parser.add_argument("--os_distro", required=True, help=f"Type of OS distro. Options: {ChromeOsSetup._OS_TYPE} | {DebianSetup._OS_TYPE} | {UbuntuOsSetup._OS_TYPE}")
-        parser.add_argument("--os_type", required=False, help=f"Type of OS. Options: {OsUtil.TYPE_OS_LINUX} | {MacOsSetup._OS_TYPE}")
-        parser.add_argument("--path_base_conf", required=False, help=self._MSG_ARG_BASE_CONF_PATH, default=self._PATH_FOLDER_CONF)
-        parser.add_argument("--path_local_conf_repo",
-                            help=self._MSG_PATH_PERMCONF_REPO,
-                            default=self._PATH_DEFAULT_PERMANENT_CONF_REPO)
-        parser.add_argument("--conf_repo_version", required=False, help="Git version of the repo e.g. 'develop'", default="develop")
-        parser.add_argument("--path_conf_dir",
-                            help=self._MSG_PATH_CONF_DIR,
-                            default=self._PATH_DEFAULT_CONFIG_CONFDIR)
-        parser.add_argument("--path_symlinks_dir", required=False, help=self._MSG_ARG_PATH_COMMON_SYMLINKS, default=self._PATH_SYMLINKS_DIR)
-        parser.add_argument("--user_id", required=False, help=self._MSG_ARG_USERID, default="")
-        parser.add_argument("--skip_setup_docker", required=False, help="Skip setup for docker", action="store_true", default=True)
-        parser.add_argument("--path_vscode_installer", required=False, help="Absolute path to the installer of VSCode.", default=self._PATH_VSCODE_INSTALLER)
-
-        args = parser.parse_args()
-        self._logger.info("args: {}".format(args))
-
-        self._logger.info("If 'user_id' is not passed, get the user id of the current process.")
-        if not args.user_id:
-            args.user_id = pwd.getpwuid(os.getuid())[0]
-
-        if not args.hostname:
-            args.hostname = os.uname()[1]
-            self._logger.warn(f"If 'hostname' is not passed, get the host name from the OS.: {args.hostname}")
-
-        return args
-
-    def main(self):
-        _args = self._cli_args()
-        # Builder pattern
-        _os_builder = None
-        if _args.os_distro == ChromeOsSetup._OS_TYPE:
-            _os_builder = ChromeOsSetup(args_in=_args)
-        elif _args.os_distro == DebianSetup._OS_TYPE:
-            _os_builder = DebianSetup(args_in=_args)
-        elif _args.os_distro == UbuntuOsSetup._OS_TYPE:
-            _os_builder = UbuntuOsSetup(args_in=_args)
-        elif _args.os_distro == MacOsSetup._OS_TYPE:
-            _os_builder = MacOsSetup(args_in=_args)
-        else:
-            raise NotImplementedError(f"Chosen OS '{_args.os}' is either not implemented or invalid.")
-
-        # Env vars per host: Bash, Emacs
-        _host_cfg = None
-        BASH_CONFIG_NAME =  ""
-        EMACS_CONFIG_NAME = ""
-        _host_cfg_brya = HostConf(_args.hostname, "130s-brya.bash", "emacs_130s-brya.el", "id_rsa_130s-brya", "id_rsa_130s-brya.pub")
-        if _args.hostname == self.HOSTNAME_P16S:
-            _host_cfg = HostConf(_args.hostname, "bashrc_130s-p16s", "emacs_130s-p16s.el", "id_rsa_130s-p16s", "id_rsa_130s-p16s.pub")
-        elif _args.hostname == self.HOSTNAME_BRYA:
-            _host_cfg = _host_cfg_brya            
-        elif _args.hostname == (self.HOSTNAME_C13_MORPH or self.HOSTNAME_ZORK16 or self.HOSTNAME_OPFYDE_RPI5):
-            _host_cfg = HostConf(_args.hostname, "130s-brya.bash", "130s-zork16.el", "id_rsa_130s-c13-morph", "id_rsa_130s-c13-morph.pub")
-        else:
-            self._logger.warning(f"'{_args.hostname=}' not matching any host. Using default config set (that of '130s-brya').")
-            _host_cfg = _host_cfg_brya
-            _host_cfg.hostname = _args.hostname
-
-        # Ref. "_MSG_ARG_BASE_CONF_PATH"
-        _conf_base_path = OsUtil.tilde_to_expand(_args.path_base_conf) if _args.path_base_conf else ""
-            
-        _os_builder.run(_args,
-                        _host_cfg,
-                        conf_repo_remote=self._URL_CONFREPO,
-                        conf_base_path=_conf_base_path)
-
-        _msg_endroll = _args.msg_endroll if _args.msg_endroll else "Setup finished."
-        self._logger.info(_msg_endroll)
-        _os_builder.listup_runtime_issues()
-        
-
-if __name__ == '__main__':
-    comp_setup = CompInitSetup()
-    comp_setup.main()
