+++ hut_10sqft_lib/README.md
@@ -1,0 +1,5 @@
+# Package `hut_10sqft`
+
+See the documentation in [doc folder](.doc/index.md).
+
+EoF
