+++ hut_10sqft_lib/package.xml
@@ -1,0 +1,17 @@
+<package format="3">
+  <name>hut_10sqft_lib</name>
+  <version>0.2.0</version>
+  <description>A library package for generic utility for setting up computer environment. NOTE: As of 2025/10, the package manifest info is mainly described in setup.py, so that the info in this package.xml file may NOT be up-to-date.  </description>
+  <maintainer email="iisaac.saito@gmail.com">Isaac Saito</maintainer>
+  <author>Isaac Saito</author>
+  <license>Apache-2.0</license>
+  <url type="bugtracker">https://github.com/130s/hut_10sqft/issues</url>
+
+  <test_depend>ament_copyright</test_depend>
+  <test_depend>ament_flake8</test_depend>
+  <test_depend>ament_pep257</test_depend>
+  <test_depend>python3-pytest</test_depend>
+  <export>
+    <build_type>ament_python</build_type>
+  </export>
+</package>
