+++ hut_10sqft_lib/pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+log_cli = True
+log_cli_level = INFO
